api/requirements.txt
drf_extensions==0.5.0
django_environ==0.4.5
Django==2.2.6
djangorestframework==3.10.3